Refactor Text component and add createStyledText utility
